+++ src/component/idatepicker-dates.component.ts
@@ -14,6 +14,7 @@ export class IDatePickerDatesComponent implements OnInit{
     @Input() dates: Array<Object>;
     @Input() selectedDate: string;
     @Input() sundayHighlight: boolean;
+    @Input() disableDays: Array<number>;
     @Output() setSelectedDate: EventEmitter<string> = new EventEmitter();
 
     ngOnInit(){
@@ -33,5 +34,10 @@ export class IDatePickerDatesComponent implements OnInit{
 
     selectDate(date: string):void{
         this.setSelectedDate.emit(date);
+    }
+
+    isDayDisabled(dayIndex: number): boolean {
+        let arr = this.disableDays;
+        return arr.includes(dayIndex);
     }
 }