+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [unreleased]
+### Changed
+- change dayLabels type from string to Array<string>
+- change onSelect event name to getSelectedDate
+
+### Added
+- add disabled attribute
+- method openDatePicker exposed
+- method closeDatePicker exposed
+- add disableDays attribute
+
 ## [14 April 2017]
 ### Changed
 - from now on css will be replaced by sass (for development)
@@ -5,9 +16,9 @@
 
 ### Added
 - bourbon (http://bourbon.io) for development
-- add attribute sundayHighlight
-- add attribute maxYear
-- add attribute minYear
+- add sundayHighlight attribute 
+- add maxYear attribute
+- add minYear attribute
 
 ### Fixed
 - reset view to calendar after click cancel on select year option